ml/requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
joblib>=1.3.0
